Stock change implemented
When an order has been completed (Right after the payment has been verified), the Stock goes down by the amount that has been ordered.
This is implemented in the CheckoutController, in the Payment ActionResult.

Also a tiny linebreak on the Product page has been removed, as I wrongfully put it there last week.