Added several things to ensure no items with 0 stock will be bought
Added an error message on the results page, and subsequently made sure no shopping cart buttons are available on books with 0 stock within the results page.

Also added an error message on the product page, and made it so that the shopping cart buttons there are unavailable when the stock of the product is 0.

So, the user can still get to the product page of any product with 0 stock, but can't add it to their shopping cart.